requirements.txt
Box2D==2.3.10
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
Farama-Notifications==0.0.6
filelock==3.29.0
fonttools==4.62.1
fsspec==2026.4.0
gymnasium==1.3.0
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.9
moviepy==2.2.1
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.2
pandas==3.0.2
pillow==11.3.0
proglog==0.1.12
pygame-ce==2.5.7
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
seaborn==0.13.2
setuptools==81.0.0
six==1.17.0
swig==4.4.1
sympy==1.14.0
torch==2.11.0
tqdm==4.67.3
typing_extensions==4.15.0
